+++ src/builders/common-tasks-builder.js
@@ -14,6 +14,7 @@ const PublishTask = require("./tasks/publish/publish");
 const UnPublishTask = require("./tasks/publish/unpublish");
 const TraceDependencies = require("./tasks/trace/trace-dependencies");
 const FixVersion = require("./tasks/version/fix-version");
+const FixDependencyVersion = require("./tasks/version/fix-dependency-version");
 const LicenseHeader = require("./tasks/license/license-header");
 
 
@@ -28,7 +29,7 @@ module.exports = {
         // Dependencies tasks
         new CleanTask("dependencies:clean","", ["dependencies:clean-test", "dependencies:clean-build"], gulp, helper, conf, project, helper.NODE_MODULES_APP);
         new CleanTask("dependencies:clean-build","", [], gulp, helper, conf, project, helper.NODE_MODULES_BUILD);
-        new CleanTask("dependencies:clean-test", "", [], gulp, helper, conf, project, [helper.NODE_MODULES_TEST, "test_report"]);
+        new CleanTask("dependencies:clean-test", "", [], gulp, helper, conf, project, helper.NODE_MODULES_TEST);
         new CleanTask("dependencies:clean-definition","", [], gulp, helper, conf, project, helper.TS_DEFINITIONS_DEPENDENCIES_PATH);
         new CleanTask("dependencies:clean-all","", ["dependencies:clean-definition", "dependencies:clean-fix"], gulp, helper, conf, project, [helper.NODE_MODULES, "test_report"]);
         new CleanDockerContainer("dependencies:clean-docker", "", [], gulp, helper, conf, project);
@@ -51,6 +52,7 @@ module.exports = {
         new TraceDependencies("dependencies:trace", "", [], gulp, helper, conf, project);
 
         new FixVersion("versions:set", "", [], gulp, helper, conf, project);
+        new FixDependencyVersion("dependency:set", "", [], gulp, helper, conf, project);
 
         new LicenseHeader("license:header", "", [], gulp, helper, conf, project);